+++ source/standalone_examples/api/omni.isaac.kaya/kaya_move.py
@@ -18,7 +18,7 @@ import numpy as np
 # TODO: changed this when asset gets converted
 my_world = World(stage_units_in_meters=0.01)
 my_kaya = my_world.scene.add(Kaya(prim_path="/World/Kaya", name="my_kaya", position=np.array([0, 0.0, 2.0])))
-my_world.scene.add_ground_plane()
+my_world.scene.add_default_ground_plane()
 my_controller = HolonomicController(name="holonomic_controller")
 my_world.reset()